Add files via upload
This code creates a graph of r vs IKSI for each subject.   Where r is the correlation between IKSI and H which measures sensitivity of typing speed to predictability of letter position.  I was trying to see if faster typists type letters at positions with low H (high predictability) faster.